Added aggregateLorPlot() to plot_comethylation_v2_functions.r Made some changes to lorQuantilePlot(), largely to improve the plot titles, legend, etc.
plot_comethylation_v2.r is a script to do quantile plots of log odds ratios and plots of aggregated log odds ratios. Currently it will stratify results by CpG islands. It is trivial to add other genomic features to stratify on. It explores the effect of the number of intervening CpGs (NIC) on the plots, also how CpG-pairs are constructed ('all' vs. 'outermost').